executor: Add logging msg
